+++ test/util.ts
@@ -1,6 +1,6 @@
 import { BN, rlp } from 'sbr-util'
 import { Block, BlockHeader } from '@sbr/block'
-import Common from '@ethereumjs/common'
+import Common from '@sbr/common'
 import Blockchain from '../src'
 const level = require('level-mem')